requirements.txt
langchain
langgraph
python-dotenv
flet
langchain-community
langchain-core
langchain-openai
langchain-anthropic
langchain-ollama
ollama
requests
tantivy
gdown
pydantic
